Moving ResourceDataUtils to src/lib/utils/
